+++ index.html
@@ -226,7 +226,7 @@
       </div>
     </section>
   </template>
- <script type="module" src="./js/functions.js"></script>
+ <!-- <script type="module" src="./js/functions.js"></script> -->
  <script type="module" src="./js/main.js"></script>
 </body>
 </html>
